# Gateway/swagger.json
{
  "swagger": "2.0",
  "consumes": [
    "application/json"
  ],
  "produces": ["application/json"],
  "info": {
    "title": "Nearby",
    "description": "",
    "version": "1.0"
  },
    "securityDefinitions": {
      "bearerAuth": {
        "type": "apiKey",
        "name": "Authorization",
        "scheme": "bearer",
        "in": "headers",
        "bearerFormat": "jwt"
      }
    },
  "security": [{
    "bearerAuth": []
  }],
  "paths": {
    "/api/auth/login": {
      "post": {
        "x-swagger-router-controller": "home",
        "operationId": "login",
        "tags": ["Auth"],
        "description": "Login with username and password",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Pet object that needs to be added to the store",
            "required": true,
            "schema": {
              "$ref": "#/definitions/login"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "examples": {
              "application/json": {
                "user": {
                  "username": "test7",
                  "email": "test7@hotmail.com",
                  "createdAt": "2021-06-09T18:48:48.897Z",
                  "updatedAt": "2021-06-09T18:48:48.897Z",
                  "id": "60c10d108a598ba073332d21"
              },
              "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjYwYzEwZDEwOGE1OThiYTA3MzMzMmQyMSIsInVzZXJuYW1lIjoidGVzdDciLCJpYXQiOjE2MjM3MDI3NDl9.-WEE5S9zUUIQJjdJaMkaJb06l9qZr5aUil7jvnMh1nE"
              }
            }
          },
          "400": {
            "description": "Invalid status value",
            "examples": {
              "application/json": {
                "message": {
                  "username": "Username is incorrect."
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "x-swagger-router-controller": "home",
        "operationId": "register",
        "tags": ["Auth"],
        "description": "Register with email, username and password",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Pet object that needs to be added to the store",
            "required": true,
            "schema": {
              "$ref": "#/definitions/register"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "examples": {
              "application/json": {
                "data": {
                  "user": {
                      "profile_photo": null,
                      "status": false,
                      "language": "en",
                      "fcm": null,
                      "about": "Available",
                      "friends": [],
                      "blocked_users": [],
                      "location": {
                          "latitude": 0,
                          "longitude": 0
                      },
                      "username": "test",
                      "email": "test@hotmail.com",
                      "createdAt": "2021-06-15T10:56:38.944Z",
                      "updatedAt": "2021-06-15T10:56:38.944Z",
                      "id": "60c88766129cfa3654cb2759"
                  }
              }
               }
            }
          },
          "400": {
            "description": "Invalid status value",
            "examples": {
              "application/json": {
                "message": "User validation failed: email: is already taken., username: is already taken."
              }
            }
          }
        }
      }
    },
    "/api/auth/update-password": {
      "post": {
        "x-swagger-router-controller": "home",
        "operationId": "update-password",
        "tags": ["Auth"],
        "description": "Update password with password and new password",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/updatePassword"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "examples": {
              "application/json": {
                "message": "Password was updated."
              }
            }
          },
          "400": {
            "description": "Invalid status value",
            "examples": {
              "application/json": {
                "message": {
                  "password": "Password is incorrect."
                }
              }
            }
          }
        }
      }
    },
    "/api/user/search": {
      "get": {
        "x-swagger-router-controller": "home",
        "operationId": "search users",
        "tags": ["Users"],
        "description": "Search users by username",
        "parameters": [
          {
            "in": "query",
            "name": "value",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "examples": {
              "application/json": [
                {
                    "profile_photo": null,
                    "status": false,
                    "language": "en",
                    "fcm": null,
                    "about": "Available",
                    "friends": [],
                    "blocked_users": [],
                    "location": {
                        "latitude": 39.933365,
                        "longitude": 32.859741
                    },
                    "username": "test10",
                    "email": "test10@hotmail.com",
                    "createdAt": "2021-06-09T18:52:11.559Z",
                    "updatedAt": "2021-06-09T18:52:11.559Z",
                    "id": "60c10ddb8a598ba073332d24"
                },
                {
                    "profile_photo": null,
                    "status": false,
                    "language": "en",
                    "fcm": null,
                    "about": "Available",
                    "friends": [],
                    "blocked_users": [],
                    "location": {
                        "latitude": 0,
                        "longitude": 0
                    },
                    "username": "test11",
                    "email": "test11@hotmail.com",
                    "createdAt": "2021-06-14T14:57:44.518Z",
                    "updatedAt": "2021-06-14T14:57:44.518Z",
                    "id": "60c76e68c5c25e54a630e5c9"
                }
            ]
            }
          },
          "404": {
            "description": "Invalid status value",
            "examples": {
              "application/json": {
                "message": "Not found"
              }
            }
          }
        }
      }
    },
    "/api/user/random": {
      "get": {
        "x-swagger-router-controller": "home",
        "operationId": "get random user",
        "tags": ["Users"],
        "description": "Get random user",
        "responses": {
          "200": {
            "description": "Successful operation",
            "examples": {
              "application/json": {
                "profile_photo": null,
                "status": false,
                "language": "en",
                "fcm": null,
                "about": "Available",
                "friends": [],
                "blocked_users": [],
                "location": {
                    "latitude": 0,
                    "longitude": 0
                },
                "username": "test5",
                "email": "test5@hotmail.com",
                "createdAt": "2021-06-09T16:36:43.520Z",
                "updatedAt": "2021-06-15T10:14:20.519Z",
                "id": "60c0ee1b8a598ba073332d1d"
              }
            }
          },
          "404": {
            "description": "Invalid token",
            "examples": {
              "application/json": {
                "message": "Not found"
              }
            }
          }
        }
      }
    },
    "/api/user/nearby": {
      "get": {
        "x-swagger-router-controller": "home",
        "operationId": "nearby user",
        "tags": ["Users"],
        "description": "Nearby users by range",
        "parameters": [
          {
            "in": "query",
            "name": "range"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "examples": {
              "application/json": [
                {
                    "profile_photo": null,
                    "status": false,
                    "language": "en",
                    "fcm": null,
                    "about": "Available",
                    "friends": [],
                    "blocked_users": [],
                    "location": {
                        "latitude": 39.933365,
                        "longitude": 32.859741
                    },
                    "_id": "60c10ddb8a598ba073332d24",
                    "username": "test10",
                    "email": "test10@hotmail.com",
                    "createdAt": "2021-06-09T18:52:11.559Z",
                    "updatedAt": "2021-06-09T18:52:11.559Z",
                    "__v": 0,
                    "id": "60c10ddb8a598ba073332d24",
                    "distance": 50
                },
                {
                    "profile_photo": null,
                    "status": false,
                    "language": "en",
                    "fcm": null,
                    "about": "Available",
                    "friends": [],
                    "blocked_users": [],
                    "location": {
                        "latitude": 0,
                        "longitude": 0
                    },
                    "_id": "60c76e68c5c25e54a630e5c9",
                    "username": "test11",
                    "email": "test11@hotmail.com",
                    "createdAt": "2021-06-14T14:57:44.518Z",
                    "updatedAt": "2021-06-14T14:57:44.518Z",
                    "__v": 0,
                    "id": "60c76e68c5c25e54a630e5c9",
                    "distance": 1000
                }
            ]
            }
          },
          "404": {
            "description": "Invalid status value",
            "examples": {
              "application/json": {
                "message": "Not found"
              }
            }
          }
        }
      }
    },
    "/api/user/profile": {
      "get": {
        "x-swagger-router-controller": "home",
        "operationId": "get profile",
        "tags": ["Profile"],
        "description": "Get current user",
        "responses": {
          "200": {
            "description": "Successful operation",
            "examples": {
              "application/json": {
                "profile_photo": null,
                "status": false,
                "language": "en",
                "fcm": null,
                "about": "Available",
                "friends": [],
                "blocked_users": [],
                "location": {
                    "latitude": 0,
                    "longitude": 0
                },
                "username": "test5",
                "email": "test5@hotmail.com",
                "createdAt": "2021-06-09T16:36:43.520Z",
                "updatedAt": "2021-06-15T10:14:20.519Z",
                "id": "60c0ee1b8a598ba073332d1d"
              }
            }
          },
          "404": {
            "description": "Invalid token",
            "examples": {
              "application/json": {
                "message": "Not found"
              }
            }
          }
        }
      },
      "delete": {
        "x-swagger-router-controller": "home",
        "operationId": "delete profile",
        "tags": ["Profile"],
        "description": "Delete current user",
        "responses": {
          "200": {
            "description": "Successful operation",
            "examples": {
              "application/json": {
                "message": "User was deleted"
              }
            }
          },
          "404": {
            "description": "Invalid token",
            "examples": {
              "application/json": {
                "message": "Not found"
              }
            }
          }
        }
      }
    },
    "/api/user/profile/update": {
      "post": {
        "x-swagger-router-controller": "home",
        "operationId": "profile update",
        "tags": ["Profile"],
        "description": "Update your profile",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/updateProfile"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "examples": {
              "application/json": {
                "message": "User was successfully updated."
              }
            }
          },
          "400": {
            "description": "Invalid status value",
            "examples": {
              "application/json": {
                
              }
            }
          }
        }
      }
    },
    "/api/user/profile/update-photo": {
      "post": {
        "x-swagger-router-controller": "home",
        "operationId": "photo update",
        "tags": ["Profile"],
        "description": "Update your photo",
        "parameters": [
          {
            "in": "formData",
            "name": "photo",
            "required": false,
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "examples": {
              "application/json": {
                "message": "Photo was successfully saved."
              }
            }
          },
          "400": {
            "description": "Invalid status value",
            "examples": {
              "application/json": {
                
              }
            }
          }
        }
      }
    },
    "/api/user/profile/friend": {
      "post": {
        "x-swagger-router-controller": "home",
        "operationId": "add friend",
        "tags": ["Profile"],
        "description": "Add friend",
        "parameters": [
          {
            "in": "body",
            "name": "id",
            "required": true,
            "schema": {
              "$ref": "#/definitions/id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "examples": {
              "application/json": [
                {
                    "profile_photo": "https://localhost:8002/uploads/2021-06-15T14:55:45.170Z_air-max-97-ayakkabısı-z3TlrlVN.jpeg",
                    "status": false,
                    "language": "en",
                    "fcm": null,
                    "about": "Available",
                    "location": {
                        "latitude": 0,
                        "longitude": 0
                    },
                    "username": "test7",
                    "email": "test7@hotmail.com",
                    "createdAt": "2021-06-09T18:48:48.921Z",
                    "updatedAt": "2021-06-15T14:55:45.182Z",
                    "id": "60c10d108a598ba073332d21"
                }
            ]
            }
          },
          "400": {
            "description": "Invalid status value",
            "examples": {
              "application/json": {
                "message": "Not found"
              }
            }
          }
        }
      }
    },
    "/api/user/profile/block": {
      "post": {
        "x-swagger-router-controller": "home",
        "operationId": "block friend",
        "tags": ["Profile"],
        "description": "Block friend",
        "parameters": [
          {
            "in": "body",
            "name": "id",
            "required": true,
            "schema": {
              "$ref": "#/definitions/id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "examples": {
              "application/json": [
                {
                    "profile_photo": "https://localhost:8002/uploads/2021-06-15T14:55:45.170Z_air-max-97-ayakkabısı-z3TlrlVN.jpeg",
                    "status": false,
                    "language": "en",
                    "fcm": null,
                    "about": "Available",
                    "location": {
                        "latitude": 0,
                        "longitude": 0
                    },
                    "username": "test7",
                    "email": "test7@hotmail.com",
                    "createdAt": "2021-06-09T18:48:48.921Z",
                    "updatedAt": "2021-06-15T14:55:45.182Z",
                    "id": "60c10d108a598ba073332d21"
                }
            ]
            }
          },
          "400": {
            "description": "Invalid status value",
            "examples": {
              "application/json": {
                "message": "Not found"
              }
            }
          }
        }
      }
    }
   
  },

  "definitions": {
    "login": {
      "properties": {
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "register": {
      "properties": {
        "username": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "updatePassword": {
      "properties": {
        "password": {
          "type": "string"
        },
        "new_password": {
          "type": "string"
        }
      }
    },
    "updateProfile": {
      "properties": {
        "username": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "language": {
          "type": "string"
        },
        "about": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "fcm": {
          "type": "string"
        },
        "location": {
          "properties": {
            "latitude": {
              "type": "number"
            },
            "longitude": {
              "type": "number"
            }
          }
        }
      }
    },
    "id": {
      "properties": {
        "id": {
          "type": "string"
        }
      }
    }
    
  }
}